[Chore #161330034] refactored flask app creation logic to app/__init.py
